Worlds Part II (Part 2) (#83)
* fix: change TotalPlayTime to ulong

* chore(game): add meta format 2004 incl extension and adapt Steam

* refactor: rename META_LENGTH contants

* chore(game): partially adapt Switch to new save format

* test: add for Worlds Part 1 and 2

* test: add Worlds Part 2 for Steam

* fix: add IsVersion553WorldsPartIIWithDifficultyTag

* fix: change description attribute for enums with names with multiple words

* chore(game): fully adapt Switch to new save format incl tests

* refactor(cli): rename some Json arguments

* test: add cli project as dependency to at least ensure it compiles

* fix(cli): add OutputEncoding for Read action

* chore(game): add detection for 5.60 incl expedition

* chore(game): update hashed technologies

+++ libNOM.io/Meta/TotalPlayTime.cs
@@ -36,7 +36,7 @@ internal static partial class TotalPlayTime
     /// </summary>
     /// <param name="json"></param>
     /// <returns></returns>
-    public static uint Get(string? json) => RegexesTotalPlayTime.Match(json)?.ToUInt32Value() ?? 0;
+    public static ulong Get(string? json) => RegexesTotalPlayTime.Match(json)?.ToUInt64Value() ?? 0;
 
     #endregion
 }
